3dt-flask-app-project/requirements.txt
astroid==2.4.2
Brotli==1.0.9
click==7.1.2
colorama==0.4.4
config==0.5.0.post0
dash==2.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==1.1.2
Flask-Compress==1.10.1
Flask-WTF==0.14.3
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.21.4
paho-mqtt==1.5.1
pandas==1.3.4
passlib==1.7.4
plotly==5.3.1
pylint==2.6.0
python-dateutil==2.8.2
pytz==2021.3
six==1.15.0
tenacity==8.0.1
toml==0.10.2
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
